Add htmap-exec Docker image and change default image to it (#153)
* move test infrastructure into tests dir

* add htmap-exec image

* updates docs
