+++ .travis.yml
@@ -15,7 +15,7 @@ matrix:
   fast_finish: true
 
 install:
-  - docker build -t htmap-test --file docker/Dockerfile --build-arg HTCONDOR_VERSION --build-arg PYTHON_VERSION=$TRAVIS_PYTHON_VERSION .
+  - docker build -t htmap-test --file tests/_inf/Dockerfile --build-arg HTCONDOR_VERSION --build-arg PYTHON_VERSION=$TRAVIS_PYTHON_VERSION .
 
 script:
-  - docker run htmap-test tests/travis.sh
+  - docker run htmap-test tests/_inf/travis.sh
