+++ docs/source/versions/v0_4_0.rst
@@ -4,6 +4,10 @@ v0.4.0
 New Features
 ------------
 
+* HTMap's default Docker image is now ``htcondor/htmap-exec``, which is produced
+  from a Dockerfile in the HTMap ``git`` repository. It is based on ``continuumio/anaconda3``,
+  with ``htmap`` itself installed as well.
+  Issue: https://github.com/htcondor/htmap/issues/152
 * Redid :class:`htmap.Map` ``stdout`` and ``stderr``.
   They are now attributes that represent sequences over the ``stdout`` and ``stderr``
   from the map components, as strings, respectively.
