+++ tests/_inf/entrypoint.sh
