+++ tests/_inf/travis.sh
